% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{areas}
\alias{areas}
\title{TSC Areas as tibble}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 146 rows and 9 columns.
}
\usage{
areas
}
\description{
This data is used in the vignette "occurrences".
}
\seealso{
Other data: 
\code{\link{areas_gj}},
\code{\link{cae_res}},
\code{\link{phys_sample}},
\code{\link{tae_res}}
}
\concept{data}
\keyword{datasets}
