% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{areas_gj}
\alias{areas_gj}
\title{TSC Areas as GeoJSON}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 146 rows and 8 columns.
}
\usage{
areas_gj
}
\description{
This data is used in the vignette "occurrences".
}
\seealso{
Other data: 
\code{\link{areas}},
\code{\link{cae_res}},
\code{\link{phys_sample}},
\code{\link{tae_res}}
}
\concept{data}
\keyword{datasets}
