% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cae_res}
\alias{cae_res}
\title{TSC Community occurrences with their Polygon representation, unparsed}
\format{
An object of class \code{tsc_api_response} of length 5.
}
\usage{
cae_res
}
\description{
This data is used in the vignette "occurrences".
}
\seealso{
Other data: 
\code{\link{areas_gj}},
\code{\link{areas}},
\code{\link{phys_sample}},
\code{\link{tae_res}}
}
\concept{data}
\keyword{datasets}
