% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineCluster.R
\name{combineCluster}
\alias{combineCluster}
\title{Combine clusters function}
\usage{
combineCluster(x, y)
}
\arguments{
\item{x}{Object of class pam. Output of getClusters.}

\item{y}{Object of class pam. Output of getClusters.}
}
\value{
Object of class 'pam'. See \code{\link[cluster]{pam.object}}
 for details
}
\description{
Combine clusters from two different clustering
}
\details{
This function combines the clusters obtained in getClusters
 with 'slope' distance and the ones obtained with Frechet distance,
  resulting in a new clustering combining both distances.
}
\examples{

data(tscR)
data <- tscR
time <- c(1,2,3)
dist_slope <- slopeDist(data, time)
dist_frechet <- frechetDistC(data, time)
slope.cluster <- getClusters(dist_slope, 3)
frechet.cluster <- getClusters(dist_frechet, 4)
comb.cluster <- combineCluster(slope.cluster, frechet.cluster)

}
\seealso{
\code{\link{getClusters}, \link{plotCluster}}
}
\author{
Fernando Pérez-Sanz (\email{fernando.perez8@um.es})

Miriam Riquelme-Pérez (\email{miriam.riquelmep@gmail.com})
}
