% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dl_mdbzip.R
\name{dl_mdbzip}
\alias{dl_mdbzip}
\title{Download, extract and open a zipped Access database from a CKAN dataset.}
\usage{
dl_mdbzip(
  resource_id,
  destdir = here::here("data"),
  dateformat = "\%m-\%d-\%Y",
  as.is = TRUE,
  verbose = get_tsc_verbose()
)
}
\arguments{
\item{resource_id}{The CKAN resource ID of a zipped Access DB.}

\item{destdir}{The local destination directory for the extracted file,
will be created if not existing. Default: \code{here::here("data")}.}

\item{dateformat}{The parameter dateformat for \code{Hmisc::mdb.get()},
default: \verb{\%Y-\%m-\%d}.}

\item{as.is}{The parameter \code{as.is} for \code{Hmisc::mdb.get()}, default: TRUE.}
}
\value{
The \code{Hmisc::mdb.get} connection.
}
\description{
The extracted file will be kept in \code{destdir}.
If a file with the expected filename already exists in \code{destdir}, this file
will be used.
To force a fresh download, remove or rename the file in \code{destdir}.
}
\examples{
\dontrun{
ckanr::ckanr_setup(
  url = Sys.getenv("CKAN_URL"), key = Sys.getenv("CKAN_API_KEY")
)
tfa_data <- dl_mdbzip("66efb68d-8f05-4bfc-af14-5d1a381d0cf2", verbose = TRUE)
}
}
\seealso{
Other ckan: 
\code{\link{ckan_works}()},
\code{\link{upload_file_to_ckan}()},
\code{\link{upload_to_ckan}()}
}
\concept{ckan}
