% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frechetDistC.R
\name{frechetDistC}
\alias{frechetDistC}
\title{frechetDistC}
\usage{
frechetDistC(x, time)
}
\arguments{
\item{x}{Numeric matrix or data.frame with trajectory values. Rows are trajectories,
columns are time or similar.}

\item{time}{Numeric vector with time data (time intervals), with equal length to columns number in x.}
}
\value{
A dist class object of size NxN, where N is rows number in the input data
}
\description{
Compute pairswise Frechet distance in a matrix of trajectories
}
\details{
This function is a C adaptation of the distFrechet code from kmlShape package for use with a matrix or
a dataframe so that the user can compute pairwise distances between all trajectories.

It is highly recommended to use this function over frechetDist because it is a lot faster.
}
\examples{

data(tscR)
time <- c(1,2,3)
dist_tscR <- frechetDistC(tscR, time)


}
\seealso{
\code{distFrechet (package kmlShape), \link{slopeDist}, \link{frechetDistR} (R and slower versión than frechetDistC).}
}
\author{
Fernando Pérez-Sanz (\code{fernando.perez8@um.es})

Miriam Riquelme Pérez (\code{miriam.riquelmep@gmail.com})
}
