% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getClusters.R
\name{getClusters}
\alias{getClusters}
\title{Generic clustering function}
\usage{
getClusters(x, k)
}
\arguments{
\item{x}{Numeric distance object obtained with dimension n x n.}

\item{k}{Numeric. Number of clusters}
}
\value{
Object of class "pam". See \code{\link[cluster]{pam.object}} for details
}
\description{
Compute clustering with pam function and a distance class object.
}
\details{
This function is a wrapper of pam (cluster). x must be a dist object obtained from frechetdist, slopedist
 or any other distance metric on condition
 that it be an object of the dist class and has dimensions nxn, where n is equal to the number of trajectories.
}
\examples{

data(tscR)
time <- c(1,2,3)
dist_tscR <- slopeDist(tscR, time)
res.cluster <- getClusters(dist_tscR, 3)


}
\seealso{
\code{\link[cluster]{pam}, \link{plotClusters}.}
}
\author{
Fernando Pérez-Sanz (\code{fernando.perez8@um.es})

Miriam Riquelme Pérez (\code{miriam.riquelmep@gmail.com})
}
