% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_http_status.R
\name{handle_http_status}
\alias{handle_http_status}
\title{Emit a specific warning messages depending on HTTP status}
\usage{
handle_http_status(response)
}
\arguments{
\item{response}{An HTTP response.}
}
\value{
The input HTTP response.
}
\description{
Status 401: Instruct how to authenticate.
Status http_error: warn and show status.
}
\seealso{
Other helpers: 
\code{\link{get_more}()},
\code{\link{tsc_msg_abort}()},
\code{\link{tsc_msg_info}()},
\code{\link{tsc_msg_noop}()},
\code{\link{tsc_msg_success}()},
\code{\link{tsc_msg_warn}()},
\code{\link{tsc_online}()}
}
\concept{helpers}
