% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineClass.R
\docType{class}
\name{imputeSenator-class}
\alias{imputeSenator-class}
\alias{show,imputeSenator-method}
\title{Class "inputeSenator"
This class represents the result imputeSenators function}
\description{
Class "inputeSenator"
This class represents the result imputeSenators function
}
\section{Slots}{

\describe{
\item{\code{data}}{Dataframe with original data}

\item{\code{senators}}{Senators data value}

\item{\code{endCluster}}{Final cluster assignment}
}}

\seealso{
\code{\link{imputeSenators}}
}
\author{
Fernando Pérez Sanz (\email{fernando.perez8@um.es})

Miriam Riquelme Pérez (\email{miriam.riquelmep@gmail.com})
}
\keyword{classes}
