% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputeSenatorToData.R
\name{imputeSenatorToData}
\alias{imputeSenatorToData}
\title{Impute clusters from senators to data}
\usage{
imputeSenatorToData(senators, clusters)
}
\arguments{
\item{senators}{List object obtained from imputeSenator function}

\item{clusters}{Pam object obtained from \link{getClusters} or
\link{combineCluster}.}
}
\value{
Object of \link{imputeSenator-class}.
}
\description{
Assign trajetories from the original data to senator clusters.
}
\details{
When it's computed a clustering over senators, it's necessary to
 assign those cluster
to original data. To do this, it's known which senator each original
trajectory belong to, therefore the final cluster of each senator is
 identified and the trajectories of that senator are assigned to its
  definitive cluster.
}
\examples{

data( tscR )
data <- tscR
time <- c( 1, 2, 3 )
senators <- imputeSenators( data, k = 100 )
senatorDist <- slopeDist( senators$senatorData, time )
sClust <- getClusters( senatorDist, k = 5 )
plotCluster( senators$senatorData, sClust, 2 )
endCluster <- imputeSenatorToData( senators, sClust )

}
\seealso{
\code{\link{plotClusterSenator}, \link{imputeSenators},
 \link{getClusters}, \link{imputeSenator-class}.}
}
\author{
Fernando Pérez-Sanz (\email{fernando.perez8@um.es})

Miriam Riquelme-Pérez (\email{miriam.riquelmep@gmail.com})
}
