% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tscr-package.R
\name{lifecycle_warning_muffler}
\alias{lifecycle_warning_muffler}
\title{Silence R CMD CHECK warning: Use lifecycle somewhere in package}
\usage{
lifecycle_warning_muffler()
}
\description{
Silence R CMD CHECK warning: Use lifecycle somewhere in package
}
\keyword{internal}
