% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotClusterSenator.R
\name{plotClusterSenator}
\alias{plotClusterSenator}
\title{Plot trajectories based on clustering}
\usage{
plotClusterSenator(x, ncluster)
}
\arguments{
\item{x}{imputesenator class object from \link{imputeSenatorToData} function.}

\item{ncluster}{When nclust = 'all', plots all trajectories and cluster
together in a single plot. If it's an integer, it draws only trajectories
 that belong to that cluster. Finally, if it is a numeric vector, it draws
  trajectories corresponding to each cluster within a subplot.}
}
\value{
Plot clustered trayectories
}
\description{
Draw trajectories and are colored based on their clusters with imputesenator
 object
}
\details{
It draws trajectories where x axis is time data and y axis
 trayectory values.
}
\examples{

data( tscR )
data <- tscR
time <- c( 1, 2, 3 )
senators <- imputeSenators( data, k = 100 )
senatorDist <- slopeDist( senators$senatorData, time )
sClust <- getClusters( senatorDist, k = 5 )
endCluster <- imputeSenatorToData( senators, sClust )
plotClusterSenator( endCluster, 'all' )
}
