% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsc_GET.R
\name{print.tsc_api_response}
\alias{print.tsc_api_response}
\title{S3 print method for 'tsc_api_response'.}
\usage{
\method{print}{tsc_api_response}(x, ...)
}
\arguments{
\item{x}{An object of class \code{tsc_api_response} as returned by
\code{\link{tsc_GET}}.}

\item{...}{Extra parameters for \code{print}}
}
\description{
Prints a short representation of data returned by
\code{\link{tsc_GET}}.
}
\seealso{
Other api: 
\code{\link{handle_http_status}()},
\code{\link{tsc_GET}()},
\code{\link{tsc_POST}()},
\code{\link{tsc_parse}()}
}
\concept{api}
