% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slopeDist.R
\name{slopeDist}
\alias{slopeDist}
\title{Pairwise slope distance}
\usage{
slopeDist(x, time)
}
\arguments{
\item{x}{Numeric matrix or data.frame with trajectory values. Rows are
trajectories, columns are time or similar. SummarizedExperiment
object can be provided for compatibility with bioconductor container (for
more information see vignette).}

\item{time}{Numeric vector with time data (time intervals), with equal
length to columns number in x.}

\item{...}{Other arguments to pass to importFromSE if _x_
is SummarizedExperiment-class.}
}
\value{
A dist class object of size NxN, where N is rows number
 in the input data
}
\description{
Compute pairwise distance based on slopes in a matrix of trajectories
}
\examples{

data(tscR)
data <- tscR
time <- c(1,2,3)
dist_tscR <- slopeDist(data, time)


}
\seealso{
\code{\link{frechetDistC} and \link{frechetDist} (R and slower
 versión than frechetDistC.), \link{importFromSE}.}
}
\author{
Fernando Pérez-Sanz (\email{fernando.perez8@um.es})

Miriam Riquelme Pérez (\email{miriam.riquelmep@gmail.com})
}
