% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tae_res}
\alias{tae_res}
\title{TSC Taxon occurrences with their Point representation, unparsed}
\format{
An object of class \code{tsc_api_response} of length 5.
}
\usage{
tae_res
}
\description{
This data is used in the vignette "occurrences".
}
\seealso{
Other data: 
\code{\link{areas_gj}},
\code{\link{areas}},
\code{\link{cae_res}},
\code{\link{phys_sample}}
}
\concept{data}
\keyword{datasets}
