% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsc_msg.R
\name{tsc_msg_abort}
\alias{tsc_msg_abort}
\title{rlang::abort() with a red error message with a cross symbol}
\usage{
tsc_msg_abort(message)
}
\arguments{
\item{message}{(chr) A message to print}
}
\description{
\lifecycle{stable}
}
\examples{
\dontrun{
tsc_msg_abort("This is an error, abort.")
}
}
\seealso{
Other helpers: 
\code{\link{get_more}()},
\code{\link{tsc_msg_info}()},
\code{\link{tsc_msg_noop}()},
\code{\link{tsc_msg_success}()},
\code{\link{tsc_msg_warn}()},
\code{\link{tsc_online}()},
\code{\link{tsc_works}()}
}
\concept{helpers}
