% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsc_msg.R
\name{tsc_msg_info}
\alias{tsc_msg_info}
\title{Print a blue info message with an info symbol}
\usage{
tsc_msg_info(message, verbose = tscr::get_tsc_verbose())
}
\arguments{
\item{message}{(chr) A message to print}

\item{verbose}{Whether to show verbose messages.
Default: \code{\link{get_tsc_verbose}}.}
}
\description{
\lifecycle{stable}
}
\examples{
tsc_msg_info("This is an info message.")
}
\seealso{
Other helpers: 
\code{\link{get_more}()},
\code{\link{tsc_msg_abort}()},
\code{\link{tsc_msg_noop}()},
\code{\link{tsc_msg_success}()},
\code{\link{tsc_msg_warn}()},
\code{\link{tsc_online}()},
\code{\link{tsc_works}()}
}
\concept{helpers}
