% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsc_msg.R
\name{tsc_msg_warn}
\alias{tsc_msg_warn}
\title{rlang::warn() with a yellow warning message with a warning symbol.}
\usage{
tsc_msg_warn(message)
}
\arguments{
\item{message}{(chr) A message to print}
}
\description{
\lifecycle{stable}
}
\examples{
\dontrun{
tsc_msg_warn("This is a warning.")
}
}
\seealso{
Other utilities: 
\code{\link{tsc_msg_abort}()},
\code{\link{tsc_msg_info}()},
\code{\link{tsc_msg_noop}()},
\code{\link{tsc_msg_success}()}
}
\concept{utilities}
