% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsc_occ_obs_post.R
\name{tsc_occ_obs_post}
\alias{tsc_occ_obs_post}
\title{Post a list of records to TSC API endpoint "occ-observation/bulk_create"}
\usage{
tsc_occ_obs_post(
  data,
  obstype = "PhysicalSample",
  chunksize = 100,
  api_url = get_tsc_api_url(),
  api_token = get_tsc_api_token(),
  verbose = get_tsc_verbose()
)
}
\arguments{
\item{data}{A tbl_df of occ-observation records}

\item{obstype}{The model type of the occ-observation model,
default: "PhysicalSample".}

\item{chunksize}{The number of records to post at a time, default: 1000.}

\item{api_url}{(character) The TSC API URL,
default \code{\link{get_tsc_api_url}}, see \code{\link{tsc_setup}}}

\item{api_token}{(character) The TSC API token,
default \code{\link{get_tsc_api_token}}, see \code{\link{tsc_setup}}}

\item{verbose}{Whether to show verbose messages.
Default: \code{\link{get_tsc_verbose}}.}
}
\description{
Post a list of records to TSC API endpoint "occ-observation/bulk_create"
}
\examples{
\dontrun{
testthat::skip_if_not(tsc_works())
data("phys_sample")
x <- tsc_occ_obs_post(phys_sample,
  obstype = "PhysicalSample",
  chunksize = 10,
  api_url = get_tsc_test_api_url(),
  api_token = get_tsc_test_api_token(),
  verbose = TRUE
)
}
}
\seealso{
Other api: 
\code{\link{handle_http_status}()},
\code{\link{print.tsc_api_response}()},
\code{\link{tsc_GET}()},
\code{\link{tsc_POST}()},
\code{\link{tsc_chunk_post}()},
\code{\link{tsc_parse}()}
}
\concept{api}
