% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsc_settings.R
\name{tsc_settings}
\alias{tsc_settings}
\alias{get_tsc_api_url}
\alias{get_tsc_api_token}
\alias{get_tsc_test_api_url}
\alias{get_tsc_test_api_token}
\alias{get_tsc_verbose}
\title{Get or set TSC settings}
\usage{
tsc_settings()

get_tsc_api_url()

get_tsc_api_token()

get_tsc_test_api_url()

get_tsc_test_api_token()

get_tsc_verbose()
}
\value{
\code{\link{tsc_settings}} prints your settings.
}
\description{
\lifecycle{stable}
}
\details{
\code{tsc_setup} sets your production and test settings.
}
\examples{
\dontrun{
tsc_settings()
}
}
\seealso{
Other helpers: 
\code{\link{get_more}()},
\code{\link{tsc_online}()}
}
\concept{helpers}
