% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsc_settings.R
\name{tsc_setup}
\alias{tsc_setup}
\title{Configure default TSC settings}
\usage{
tsc_setup(
  api_url = NULL,
  api_token = NULL,
  test_api_url = NULL,
  test_api_token = NULL,
  verbose = NULL
)
}
\arguments{
\item{api_url}{A TSC API URL (optional),
default: \code{\link{get_tsc_api_url}}.}

\item{api_token}{A TSC API token, leading with "Token "}

\item{test_api_url}{A TSC API URL for tests (optional),
default: \code{\link{get_tsc_test_api_url}}}

\item{test_api_token}{A TSC API token for tests,
leading with "Token " (character).
Default: \code{\link{get_tsc_api_token}}.}

\item{verbose}{Verbosity for tscr.}
}
\description{
\lifecycle{stable}
}
\details{
\code{tscr_setup} sets TSC API connection details. \code{tscr}'s
functions default to use the default URL, but require an API Token to work.
}
\examples{
# TSC users can access their TSC API Token on their profile page on TSC.
# Not specifying the default TSC API URL will reset the TSC URL to its
# default "https://tsc.dbca.wa.gov.au/api/1/":
tsc_setup(api_token = "Token xxx")
}
