% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsc_works.R
\name{tsc_works}
\alias{tsc_works}
\title{Check whether API token is set and TSC is online}
\usage{
tsc_works(
  api_url = get_tsc_api_url(),
  api_token = get_tsc_api_token(),
  api_un = get_tsc_api_un(),
  api_pw = get_tsc_api_pw()
)
}
\arguments{
\item{api_url}{(character) The TSC API URL,
default \code{\link{get_tsc_api_url}}, see \code{\link{tsc_setup}}}

\item{api_token}{(character) The TSC API token,
default \code{\link{get_tsc_api_token}}, see \code{\link{tsc_setup}}}
}
\description{
\lifecycle{stable}
}
\seealso{
Other helpers: 
\code{\link{get_more}()},
\code{\link{tsc_msg_abort}()},
\code{\link{tsc_msg_info}()},
\code{\link{tsc_msg_noop}()},
\code{\link{tsc_msg_success}()},
\code{\link{tsc_msg_warn}()},
\code{\link{tsc_online}()}
}
\concept{helpers}
