% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tscR}
\alias{tscR}
\title{Dummy trajectories data}
\format{A data frame 300 rows and 3 columns:
\describe{
  \item{T1}{time interval}
  \item{T2}{time interval}
  \item{T3}{time interval}
}}
\source{
Simulated data
}
\usage{
tscR
}
\description{
A dataset containing 300 tajectories and 3 time points
}
\keyword{datasets}
