% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_file_to_ckan.R
\name{upload_file_to_ckan}
\alias{upload_file_to_ckan}
\title{Upload a file to an existing CKAN resource ID or skip if file is missing}
\usage{
upload_file_to_ckan(rid, fn, verbose = get_tsc_verbose())
}
\arguments{
\item{rid}{An existing CKAN resource ID.}

\item{fn}{The file to upload as the CKAN file resource.}

\item{verbose}{Whether to show verbose messages.
Default: \code{\link{get_tsc_verbose}}.}
}
\value{
NULL if the file is missing, or the resource ID.
}
\description{
This function can be used to update an existing CKAN file resource with a
compiled Rmd workbook (i.e a HTML file).
}
\details{
This function is a wrapper around \code{ckanr::resource_update()} with contingencies
for missing files.
}
\examples{
\dontrun{
ckanr::ckanr_setup(
  url = Sys.getenv("CKAN_URL"), key = Sys.getenv("CKAN_API_KEY")
)
upload_file_to_ckan("1b5e6401-82fe-4728-83be-f3f8ea4b5876", "EDA_flora.html")
}
}
\seealso{
Other ckan: 
\code{\link{ckan_works}()},
\code{\link{dl_mdbzip}()},
\code{\link{upload_to_ckan}()}
}
\concept{ckan}
