% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_to_ckan.R
\name{upload_to_ckan}
\alias{upload_to_ckan}
\title{Create or update a CKAN resource.}
\usage{
upload_to_ckan(
  data,
  resource_title,
  dataset_id,
  resource_id = NULL,
  verbose = get_tsc_verbose()
)
}
\arguments{
\item{data}{A data frame to write to disk.}

\item{resource_title}{A CKAN resource title.}

\item{dataset_id}{A CKAN dataset (package) ID.}

\item{resource_id}{A CKAN resource ID, default: NULL.}

\item{verbose}{Whether to show verbose messages.
Default: \code{\link{get_tsc_verbose}}.}
}
\value{
The resource ID of the created or updated resource.
}
\description{
Create or update a CKAN resource.
}
\details{
The data will be written to CSV in a directory \verb{data/} with the
resource title in snake_case.
If no resource ID is given, a resource will be created.
The resource ID is returned in either case.
}
\examples{
\dontrun{
ckanr::ckanr_setup(
  url = Sys.getenv("CKAN_URL"), key = Sys.getenv("CKAN_API_KEY")
)
d <- ckanr::package_show("threatened-ecological-communities-database")

# Run this once to create resource and retrieve resource ID
upload_to_ckan(a_tibble, "Resource title", d$id, resource_id = NULL)
# returns "502c74d7-32be-453f-aff6-c50aedd3deed" - paste into resource_id

# Re-run this to update resource with new data
upload_to_ckan(a_tibble, "Resource title", d$id,
  resource_id = "502c74d7-32be-453f-aff6-c50aedd3deed"
)
}
}
\seealso{
Other ckan: 
\code{\link{ckan_works}()},
\code{\link{dl_mdbzip}()},
\code{\link{upload_file_to_ckan}()}
}
\concept{ckan}
