% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upsert_geojson.R
\name{upsert_geojson}
\alias{upsert_geojson}
\title{Upsert GeoJSON into TSC API endpoints}
\usage{
upsert_geojson(
  gj_featurecollection,
  serializer = "names",
  chunksize = 1000,
  api_url = get_tsc_api_url(),
  api_token = get_tsc_api_token(),
  verbose = get_tsc_verbose()
)
}
\arguments{
\item{gj_featurecollection}{(list) A GeoJSON featurecollection as list}

\item{serializer}{(character) TSC API serializer name (required)
Possible values see \code{https://tsc.dbca.wa.gov.au/api/1/}}

\item{chunksize}{(int) The number of features to upload simultaneously,
default: 1000.}

\item{api_url}{(character) The TSC API URL,
default \code{\link{get_tsc_api_url}}, see \code{\link{tsc_setup}}}

\item{api_token}{(character) The TSC API token,
default \code{\link{get_tsc_api_token}}, see \code{\link{tsc_setup}}}

\item{verbose}{Whether to show verbose messages.
Default: \code{\link{get_tsc_verbose}}.}
}
\description{
Upsert GeoJSON into TSC API endpoints
}
\seealso{
Other wacensus: 
\code{\link{gs_getWFS}()}
}
\concept{wacensus}
