% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frechetDist.R
\name{frechetDist}
\alias{frechetDist}
\title{Pairwise Frechet distance}
\usage{
frechetDist(x, time)
}
\arguments{
\item{x}{Numeric matrix or data.frame with trajectory values.
Rows are trajectories, columns are time or similar.}

\item{time}{Numeric vector with time data (time intervals),
with equal length to columns number in x.}
}
\value{
A dist class object of size NxN, where N is rows number
 in the input data
}
\description{
Compute pair-wise Frechet distance in a matrix of trajectories
}
\details{
This function is a wrapper of the distFrechet code from
 kmlShape package for use with a matrix or
a data.frame so that the user can compute
 pairwise distances between all trajectories.
}
\examples{

data(tscR)
data <- tscR
time <- c(1,2,3)
dist_tscR <- frechetDist(data, time)


}
\seealso{
\code{\link[kmlShape]{distFrechet} (package kmlShape),
 \link{slopeDist}, \link{frechetDistC}
  (C and faster versión than frechetDist).}
}
\author{
Fernando Pérez-Sanz (\email{fernando.perez8@um.es})

Miriam Riquelme-Pérez (\email{miriam.riquelmep@gmail.com})
}
