% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_http_status.R
\name{handle_http_status}
\alias{handle_http_status}
\title{Emit a specific warning messages depending on HTTP status}
\usage{
handle_http_status(response, verbose = tscr::get_tsc_verbose())
}
\arguments{
\item{response}{An HTTP response.}

\item{verbose}{Whether to show verbose messages.
Default: \code{\link{get_tsc_verbose}}.}
}
\description{
Status 401: Instruct how to authenticate.
Status http_error: warn and show status.
}
\seealso{
Other api: 
\code{\link{print.tsc_api_response}()},
\code{\link{tsc_GET}()},
\code{\link{tsc_POST}()},
\code{\link{tsc_parse}()}
}
\concept{api}
