% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{phys_sample}
\alias{phys_sample}
\title{TSC Occurrence Observations of type "Physical Sample"}
\format{
An object of class \code{data.frame} with 100 rows and 7 columns.
}
\usage{
phys_sample
}
\description{
This data is used in the examples and tests for \code{tsc_occ_obs_post}.
}
\concept{data}
\keyword{datasets}
