% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCluster.R
\name{plotCluster}
\alias{plotCluster}
\title{Plot trajectories based on clustering}
\usage{
plotCluster(data, clust, ncluster)
}
\arguments{
\item{data}{Numeric data frame or matrix with de original data.}

\item{clust}{Object of class pam or partition obtained from getClusters output.}

\item{ncluster}{When nclust = 'all', plots all trajectories and cluster together in a single plot.
If it's an integer, it draws only trajectories that belong to that cluster. Finally, if
it is a numeric vector, it draws trajectories corresponding to each cluster within a subplot.}
}
\value{
Plot clustered trayectories
}
\description{
Draw trajectories and are colored based on their clusters
}
\details{
It draws trajectories where x axis is time data and y axis trayectory values.
}
\examples{

data(tscR)
data <- tscR
time <- c(1,2,3)
fdist <- frechetDistC(data, time)
fclust <- getClusters(fdist, 3)
plotCluster(data, fclust, 'all')


}
\seealso{
\code{\link[graphics]{matplot}, \link{plotClusterSenator}}.
}
\author{
Fernando Pérez-Sanz (\email{fernando.perez8@um.es})

Miriam Riquelme-Pérez (\email{miriam.riquelmep@gmail.com})
}
