% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slopeDist.R
\name{slopeDist}
\alias{slopeDist}
\title{slopeDist}
\usage{
slopeDist(x, time)
}
\arguments{
\item{x}{Numeric matrix or data.frame with trajectory values. Rows are trajectories,
columns are time or similar.}

\item{time}{Numeric vector with time data (time intervals), with equal length to columns number in x.}
}
\value{
A dist class object of size NxN, where N is rows number in the input data
}
\description{
Compute pairswise distance based on slopes in a matrix of trajectories
}
\examples{

data(tscR)
time <- c(1,2,3)
dist_tscR <- slopeDist(tscR, time)


}
\seealso{
\code{\link{frechetDistC} and \link{frechetDist} (R and slower versión than frechetDistC.)}
}
\author{
Fernando Pérez-Sanz (\code{fernando.perez8@um.es})

Miriam Riquelme Pérez (\code{miriam.riquelmep@gmail.com})
}
