% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsc_msg.R
\name{tsc_msg_info}
\alias{tsc_msg_info}
\title{Print a blue info message with an info symbol.}
\usage{
tsc_msg_info(message)
}
\arguments{
\item{message}{(chr) A message to print}
}
\description{
\lifecycle{stable}
}
\examples{
tsc_msg_info("This is an info message.")
}
\seealso{
Other utilities: 
\code{\link{tsc_msg_abort}()},
\code{\link{tsc_msg_noop}()},
\code{\link{tsc_msg_success}()},
\code{\link{tsc_msg_warn}()}
}
\concept{utilities}
