% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tscr-package.R
\docType{package}
\name{tscr-package}
\alias{tscr}
\alias{tscr-package}
\title{tscr: Threatened Species and Communities Data Export and Analysis}
\description{
TSC is DBCA's Threatened Species and Communities Data Portal.
  TSC provides access to its data through a RESTful API. 
  tscr provides helpers and working examples of getting data in and out of TSC's
  API, as well as starting points for data visualisations and analyses of TSC 
  data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dbca-wa/tscr}
  \item \url{https://dbca-wa.github.io/tscr/}
  \item Report bugs at \url{https://github.com/dbca-wa/tscr/issues/new/choose}
}

}
\author{
\strong{Maintainer}: Florian W. Mayer \email{Florian.Mayer@dbca.wa.gov.au} (\href{https://orcid.org/0000-0003-4269-4242}{ORCID})

Other contributors:
\itemize{
  \item DBCA [copyright holder, funder]
}

}
\keyword{internal}
